Crate swc_css_visit

Modules

fields

This module contains enums representing fields of each types

Structs

All

Visit all children nodes. This converts VisitAll to Visit. The type parameter V should implement VisitAll and All<V> implements Visit.

Enums

AstParentKind

AstParentNodeRef

Traits

Fold

FoldAstPathpath

FoldWith

A utility trait implemented for ast nodes, and allow to visit them with a visitor.

FoldWithPathpath

A utility trait implemented for ast nodes, and allow to visit them with a visitor.

Node

Visitable nodes.

Visit

VisitAll

VisitAllWith

A utility trait implemented for ast nodes, and allow to visit them with a visitor.

VisitAstPathpath

VisitMut

VisitMutAstPathpath

VisitMutWith

A utility trait implemented for ast nodes, and allow to visit them with a visitor.

VisitMutWithPathpath

A utility trait implemented for ast nodes, and allow to visit them with a visitor.

VisitWith

A utility trait implemented for ast nodes, and allow to visit them with a visitor.

VisitWithPathpath

A utility trait implemented for ast nodes, and allow to visit them with a visitor.

Functions

fold_absolute_color_base

Visits children of the nodes with the given visitor.

fold_alpha_value

Visits children of the nodes with the given visitor.

fold_an_plus_b

Visits children of the nodes with the given visitor.

fold_an_plus_b_notation

Visits children of the nodes with the given visitor.

fold_angle

Visits children of the nodes with the given visitor.

fold_angle_percentage

Visits children of the nodes with the given visitor.

fold_any_namespace

Visits children of the nodes with the given visitor.

fold_at_rule

Visits children of the nodes with the given visitor.

fold_at_rule_name

Visits children of the nodes with the given visitor.

fold_at_rule_prelude

Visits children of the nodes with the given visitor.

fold_atom

Visits children of the nodes with the given visitor.

fold_attribute_selector

Visits children of the nodes with the given visitor.

fold_attribute_selector_matcher

Visits children of the nodes with the given visitor.

fold_attribute_selector_matcher_value

Visits children of the nodes with the given visitor.

fold_attribute_selector_modifier

Visits children of the nodes with the given visitor.

fold_attribute_selector_value

Visits children of the nodes with the given visitor.

fold_calc_operator

Visits children of the nodes with the given visitor.

fold_calc_operator_type

Visits children of the nodes with the given visitor.

fold_calc_product

Visits children of the nodes with the given visitor.

fold_calc_product_or_operator

Visits children of the nodes with the given visitor.

fold_calc_product_or_operators

Visits children of the nodes with the given visitor.

fold_calc_sum

Visits children of the nodes with the given visitor.

fold_calc_value

Visits children of the nodes with the given visitor.

fold_calc_value_or_operator

Visits children of the nodes with the given visitor.

fold_calc_value_or_operators

Visits children of the nodes with the given visitor.

fold_class_selector

Visits children of the nodes with the given visitor.

fold_cmyk_component

Visits children of the nodes with the given visitor.

fold_color

Visits children of the nodes with the given visitor.

fold_color_profile_name

Visits children of the nodes with the given visitor.

fold_combinator

Visits children of the nodes with the given visitor.

fold_combinator_value

Visits children of the nodes with the given visitor.

fold_complex_selector

Visits children of the nodes with the given visitor.

fold_complex_selector_children

Visits children of the nodes with the given visitor.

fold_complex_selector_children_vec

Visits children of the nodes with the given visitor.

fold_complex_selectors

Visits children of the nodes with the given visitor.

fold_component_value

Visits children of the nodes with the given visitor.

fold_component_values

Visits children of the nodes with the given visitor.

fold_compound_selector

Visits children of the nodes with the given visitor.

fold_compound_selector_list

Visits children of the nodes with the given visitor.

fold_compound_selectors

Visits children of the nodes with the given visitor.

fold_container_condition

Visits children of the nodes with the given visitor.

fold_container_name

Visits children of the nodes with the given visitor.

fold_container_query

Visits children of the nodes with the given visitor.

fold_container_query_and

Visits children of the nodes with the given visitor.

fold_container_query_not

Visits children of the nodes with the given visitor.

fold_container_query_or

Visits children of the nodes with the given visitor.

fold_container_query_type

Visits children of the nodes with the given visitor.

fold_container_query_types

Visits children of the nodes with the given visitor.

fold_custom_highlight_name

Visits children of the nodes with the given visitor.

fold_custom_ident

Visits children of the nodes with the given visitor.

fold_custom_idents

Visits children of the nodes with the given visitor.

fold_custom_media_query

Visits children of the nodes with the given visitor.

fold_custom_media_query_media_type

Visits children of the nodes with the given visitor.

fold_custom_property_name

Visits children of the nodes with the given visitor.

fold_dashed_ident

Visits children of the nodes with the given visitor.

fold_declaration

Visits children of the nodes with the given visitor.

fold_declaration_name

Visits children of the nodes with the given visitor.

fold_declaration_or_at_rule

Visits children of the nodes with the given visitor.

fold_delimiter

Visits children of the nodes with the given visitor.

fold_delimiter_value

Visits children of the nodes with the given visitor.

fold_dimension

Visits children of the nodes with the given visitor.

fold_document_prelude

Visits children of the nodes with the given visitor.

fold_document_prelude_matching_function

Visits children of the nodes with the given visitor.

fold_document_prelude_matching_functions

Visits children of the nodes with the given visitor.

fold_extension_name

Visits children of the nodes with the given visitor.

fold_family_name

Visits children of the nodes with the given visitor.

fold_family_names

Visits children of the nodes with the given visitor.

fold_flex

Visits children of the nodes with the given visitor.

fold_font_feature_values_prelude

Visits children of the nodes with the given visitor.

fold_forgiving_complex_selector

Visits children of the nodes with the given visitor.

fold_forgiving_complex_selectors

Visits children of the nodes with the given visitor.

fold_forgiving_relative_selector

Visits children of the nodes with the given visitor.

fold_forgiving_relative_selector_list

Visits children of the nodes with the given visitor.

fold_forgiving_relative_selectors

Visits children of the nodes with the given visitor.

fold_forgiving_selector_list

Visits children of the nodes with the given visitor.

fold_frequency

Visits children of the nodes with the given visitor.

fold_frequency_percentage

Visits children of the nodes with the given visitor.

fold_function

Visits children of the nodes with the given visitor.

fold_function_name

Visits children of the nodes with the given visitor.

fold_general_enclosed

Visits children of the nodes with the given visitor.

fold_hex_color

Visits children of the nodes with the given visitor.

fold_hue

Visits children of the nodes with the given visitor.

fold_i_32

Visits children of the nodes with the given visitor.

fold_id_selector

Visits children of the nodes with the given visitor.

fold_ident

Visits children of the nodes with the given visitor.

fold_idents

Visits children of the nodes with the given visitor.

fold_import_conditions

Visits children of the nodes with the given visitor.

fold_import_href

Visits children of the nodes with the given visitor.

fold_import_layer_name

Visits children of the nodes with the given visitor.

fold_import_prelude

Visits children of the nodes with the given visitor.

fold_important_flag

Visits children of the nodes with the given visitor.

fold_integer

Visits children of the nodes with the given visitor.

fold_keyframe_block

Visits children of the nodes with the given visitor.

fold_keyframe_selector

Visits children of the nodes with the given visitor.

fold_keyframe_selectors

Visits children of the nodes with the given visitor.

fold_keyframes_name

Visits children of the nodes with the given visitor.

fold_keyframes_pseudo_function

Visits children of the nodes with the given visitor.

fold_keyframes_pseudo_prefix

Visits children of the nodes with the given visitor.

fold_layer_name

Visits children of the nodes with the given visitor.

fold_layer_name_list

Visits children of the nodes with the given visitor.

fold_layer_names

Visits children of the nodes with the given visitor.

fold_layer_prelude

Visits children of the nodes with the given visitor.

fold_length

Visits children of the nodes with the given visitor.

fold_length_percentage

Visits children of the nodes with the given visitor.

fold_list_of_component_values

Visits children of the nodes with the given visitor.

fold_media_and

Visits children of the nodes with the given visitor.

fold_media_condition

Visits children of the nodes with the given visitor.

fold_media_condition_all_type

Visits children of the nodes with the given visitor.

fold_media_condition_all_types

Visits children of the nodes with the given visitor.

fold_media_condition_type

Visits children of the nodes with the given visitor.

fold_media_condition_without_or

Visits children of the nodes with the given visitor.

fold_media_condition_without_or_type

Visits children of the nodes with the given visitor.

fold_media_condition_without_or_types

Visits children of the nodes with the given visitor.

fold_media_feature

Visits children of the nodes with the given visitor.

fold_media_feature_boolean

Visits children of the nodes with the given visitor.

fold_media_feature_name

Visits children of the nodes with the given visitor.

fold_media_feature_plain

Visits children of the nodes with the given visitor.

fold_media_feature_range

Visits children of the nodes with the given visitor.

fold_media_feature_range_comparison

Visits children of the nodes with the given visitor.

fold_media_feature_range_interval

Visits children of the nodes with the given visitor.

fold_media_feature_value

Visits children of the nodes with the given visitor.

fold_media_in_parens

Visits children of the nodes with the given visitor.

fold_media_not

Visits children of the nodes with the given visitor.

fold_media_or

Visits children of the nodes with the given visitor.

fold_media_queries

Visits children of the nodes with the given visitor.

fold_media_query

Visits children of the nodes with the given visitor.

fold_media_query_list

Visits children of the nodes with the given visitor.

fold_media_type

Visits children of the nodes with the given visitor.

fold_named_namespace

Visits children of the nodes with the given visitor.

fold_namespace

Visits children of the nodes with the given visitor.

fold_namespace_prefix

Visits children of the nodes with the given visitor.

fold_namespace_prelude

Visits children of the nodes with the given visitor.

fold_namespace_prelude_uri

Visits children of the nodes with the given visitor.

fold_nesting_selector

Visits children of the nodes with the given visitor.

fold_number

Visits children of the nodes with the given visitor.

fold_opt_at_rule_prelude

Visits children of the nodes with the given visitor.

fold_opt_atom

Visits children of the nodes with the given visitor.

fold_opt_attribute_selector_matcher

Visits children of the nodes with the given visitor.

fold_opt_attribute_selector_modifier

Visits children of the nodes with the given visitor.

fold_opt_attribute_selector_value

Visits children of the nodes with the given visitor.

fold_opt_combinator

Visits children of the nodes with the given visitor.

fold_opt_container_name

Visits children of the nodes with the given visitor.

fold_opt_forgiving_selector_list

Visits children of the nodes with the given visitor.

fold_opt_function

Visits children of the nodes with the given visitor.

fold_opt_i_32

Visits children of the nodes with the given visitor.

fold_opt_ident

Visits children of the nodes with the given visitor.

fold_opt_import_conditions

Visits children of the nodes with the given visitor.

fold_opt_import_layer_name

Visits children of the nodes with the given visitor.

fold_opt_important_flag

Visits children of the nodes with the given visitor.

fold_opt_media_condition_type

Visits children of the nodes with the given visitor.

fold_opt_media_query_list

Visits children of the nodes with the given visitor.

fold_opt_media_type

Visits children of the nodes with the given visitor.

fold_opt_namespace

Visits children of the nodes with the given visitor.

fold_opt_namespace_prefix

Visits children of the nodes with the given visitor.

fold_opt_nesting_selector

Visits children of the nodes with the given visitor.

fold_opt_number

Visits children of the nodes with the given visitor.

fold_opt_page_selector_pseudos

Visits children of the nodes with the given visitor.

fold_opt_page_selector_type

Visits children of the nodes with the given visitor.

fold_opt_pseudo_class_selector_children_vec

Visits children of the nodes with the given visitor.

fold_opt_pseudo_element_selector_children_vec

Visits children of the nodes with the given visitor.

fold_opt_simple_block

Visits children of the nodes with the given visitor.

fold_opt_type_selector

Visits children of the nodes with the given visitor.

fold_opt_url_modifiers

Visits children of the nodes with the given visitor.

fold_opt_url_value

Visits children of the nodes with the given visitor.

fold_page_selector

Visits children of the nodes with the given visitor.

fold_page_selector_list

Visits children of the nodes with the given visitor.

fold_page_selector_pseudo

Visits children of the nodes with the given visitor.

fold_page_selector_pseudos

Visits children of the nodes with the given visitor.

fold_page_selector_type

Visits children of the nodes with the given visitor.

fold_page_selectors

Visits children of the nodes with the given visitor.

fold_percentage

Visits children of the nodes with the given visitor.

fold_pseudo_class_selector

Visits children of the nodes with the given visitor.

fold_pseudo_class_selector_children

Visits children of the nodes with the given visitor.

fold_pseudo_class_selector_children_vec

Visits children of the nodes with the given visitor.

fold_pseudo_element_selector

Visits children of the nodes with the given visitor.

fold_pseudo_element_selector_children

Visits children of the nodes with the given visitor.

fold_pseudo_element_selector_children_vec

Visits children of the nodes with the given visitor.

fold_qualified_rule

Visits children of the nodes with the given visitor.

fold_qualified_rule_prelude

Visits children of the nodes with the given visitor.

fold_query_in_parens

Visits children of the nodes with the given visitor.

fold_ratio

Visits children of the nodes with the given visitor.

fold_relative_selector

Visits children of the nodes with the given visitor.

fold_relative_selector_list

Visits children of the nodes with the given visitor.

fold_relative_selectors

Visits children of the nodes with the given visitor.

fold_resolution

Visits children of the nodes with the given visitor.

fold_rule

Visits children of the nodes with the given visitor.

fold_rules

Visits children of the nodes with the given visitor.

fold_scope_range

Visits children of the nodes with the given visitor.

fold_selector_list

Visits children of the nodes with the given visitor.

fold_sequence_of_custom_idents

Visits children of the nodes with the given visitor.

fold_simple_block

Visits children of the nodes with the given visitor.

fold_size_feature

Visits children of the nodes with the given visitor.

fold_size_feature_boolean

Visits children of the nodes with the given visitor.

fold_size_feature_name

Visits children of the nodes with the given visitor.

fold_size_feature_plain

Visits children of the nodes with the given visitor.

fold_size_feature_range

Visits children of the nodes with the given visitor.

fold_size_feature_range_comparison

Visits children of the nodes with the given visitor.

fold_size_feature_range_interval

Visits children of the nodes with the given visitor.

fold_size_feature_value

Visits children of the nodes with the given visitor.

fold_span

Visits children of the nodes with the given visitor.

fold_str

Visits children of the nodes with the given visitor.

fold_style_block

Visits children of the nodes with the given visitor.

fold_stylesheet

Visits children of the nodes with the given visitor.

fold_subclass_selector

Visits children of the nodes with the given visitor.

fold_subclass_selectors

Visits children of the nodes with the given visitor.

fold_supports_and

Visits children of the nodes with the given visitor.

fold_supports_condition

Visits children of the nodes with the given visitor.

fold_supports_condition_type

Visits children of the nodes with the given visitor.

fold_supports_condition_types

Visits children of the nodes with the given visitor.

fold_supports_feature

Visits children of the nodes with the given visitor.

fold_supports_in_parens

Visits children of the nodes with the given visitor.

fold_supports_not

Visits children of the nodes with the given visitor.

fold_supports_or

Visits children of the nodes with the given visitor.

fold_tag_name_selector

Visits children of the nodes with the given visitor.

fold_time

Visits children of the nodes with the given visitor.

fold_time_percentage

Visits children of the nodes with the given visitor.

fold_token

Visits children of the nodes with the given visitor.

fold_token_and_span

Visits children of the nodes with the given visitor.

fold_type_selector

Visits children of the nodes with the given visitor.

fold_unicode_range

Visits children of the nodes with the given visitor.

fold_universal_selector

Visits children of the nodes with the given visitor.

fold_unknown_dimension

Visits children of the nodes with the given visitor.

fold_url

Visits children of the nodes with the given visitor.

fold_url_modifier

Visits children of the nodes with the given visitor.

fold_url_modifiers

Visits children of the nodes with the given visitor.

fold_url_value

Visits children of the nodes with the given visitor.

fold_url_value_raw

Visits children of the nodes with the given visitor.

fold_wq_name

Visits children of the nodes with the given visitor.

visit_absolute_color_base

Visits children of the nodes with the given visitor.

visit_alpha_value

Visits children of the nodes with the given visitor.

visit_an_plus_b

Visits children of the nodes with the given visitor.

visit_an_plus_b_notation

Visits children of the nodes with the given visitor.

visit_angle

Visits children of the nodes with the given visitor.

visit_angle_percentage

Visits children of the nodes with the given visitor.

visit_any_namespace

Visits children of the nodes with the given visitor.

visit_at_rule

Visits children of the nodes with the given visitor.

visit_at_rule_name

Visits children of the nodes with the given visitor.

visit_at_rule_prelude

Visits children of the nodes with the given visitor.

visit_atom

Visits children of the nodes with the given visitor.

visit_attribute_selector

Visits children of the nodes with the given visitor.

visit_attribute_selector_matcher

Visits children of the nodes with the given visitor.

visit_attribute_selector_matcher_value

Visits children of the nodes with the given visitor.

visit_attribute_selector_modifier

Visits children of the nodes with the given visitor.

visit_attribute_selector_value

Visits children of the nodes with the given visitor.

visit_calc_operator

Visits children of the nodes with the given visitor.

visit_calc_operator_type

Visits children of the nodes with the given visitor.

visit_calc_product

Visits children of the nodes with the given visitor.

visit_calc_product_or_operator

Visits children of the nodes with the given visitor.

visit_calc_product_or_operators

Visits children of the nodes with the given visitor.

visit_calc_sum

Visits children of the nodes with the given visitor.

visit_calc_value

Visits children of the nodes with the given visitor.

visit_calc_value_or_operator

Visits children of the nodes with the given visitor.

visit_calc_value_or_operators

Visits children of the nodes with the given visitor.

visit_class_selector

Visits children of the nodes with the given visitor.

visit_cmyk_component

Visits children of the nodes with the given visitor.

visit_color

Visits children of the nodes with the given visitor.

visit_color_profile_name

Visits children of the nodes with the given visitor.

visit_combinator

Visits children of the nodes with the given visitor.

visit_combinator_value

Visits children of the nodes with the given visitor.

visit_complex_selector

Visits children of the nodes with the given visitor.

visit_complex_selector_children

Visits children of the nodes with the given visitor.

visit_complex_selector_children_vec

Visits children of the nodes with the given visitor.

visit_complex_selectors

Visits children of the nodes with the given visitor.

visit_component_value

Visits children of the nodes with the given visitor.

visit_component_values

Visits children of the nodes with the given visitor.

visit_compound_selector

Visits children of the nodes with the given visitor.

visit_compound_selector_list

Visits children of the nodes with the given visitor.

visit_compound_selectors

Visits children of the nodes with the given visitor.

visit_container_condition

Visits children of the nodes with the given visitor.

visit_container_name

Visits children of the nodes with the given visitor.

visit_container_query

Visits children of the nodes with the given visitor.

visit_container_query_and

Visits children of the nodes with the given visitor.

visit_container_query_not

Visits children of the nodes with the given visitor.

visit_container_query_or

Visits children of the nodes with the given visitor.

visit_container_query_type

Visits children of the nodes with the given visitor.

visit_container_query_types

Visits children of the nodes with the given visitor.

visit_custom_highlight_name

Visits children of the nodes with the given visitor.

visit_custom_ident

Visits children of the nodes with the given visitor.

visit_custom_idents

Visits children of the nodes with the given visitor.

visit_custom_media_query

Visits children of the nodes with the given visitor.

visit_custom_media_query_media_type

Visits children of the nodes with the given visitor.

visit_custom_property_name

Visits children of the nodes with the given visitor.

visit_dashed_ident

Visits children of the nodes with the given visitor.

visit_declaration

Visits children of the nodes with the given visitor.

visit_declaration_name

Visits children of the nodes with the given visitor.

visit_declaration_or_at_rule

Visits children of the nodes with the given visitor.

visit_delimiter

Visits children of the nodes with the given visitor.

visit_delimiter_value

Visits children of the nodes with the given visitor.

visit_dimension

Visits children of the nodes with the given visitor.

visit_document_prelude

Visits children of the nodes with the given visitor.

visit_document_prelude_matching_function

Visits children of the nodes with the given visitor.

visit_document_prelude_matching_functions

Visits children of the nodes with the given visitor.

visit_extension_name

Visits children of the nodes with the given visitor.

visit_family_name

Visits children of the nodes with the given visitor.

visit_family_names

Visits children of the nodes with the given visitor.

visit_flex

Visits children of the nodes with the given visitor.

visit_font_feature_values_prelude

Visits children of the nodes with the given visitor.

visit_forgiving_complex_selector

Visits children of the nodes with the given visitor.

visit_forgiving_complex_selectors

Visits children of the nodes with the given visitor.

visit_forgiving_relative_selector

Visits children of the nodes with the given visitor.

visit_forgiving_relative_selector_list

Visits children of the nodes with the given visitor.

visit_forgiving_relative_selectors

Visits children of the nodes with the given visitor.

visit_forgiving_selector_list

Visits children of the nodes with the given visitor.

visit_frequency

Visits children of the nodes with the given visitor.

visit_frequency_percentage

Visits children of the nodes with the given visitor.

visit_function

Visits children of the nodes with the given visitor.

visit_function_name

Visits children of the nodes with the given visitor.

visit_general_enclosed

Visits children of the nodes with the given visitor.

visit_hex_color

Visits children of the nodes with the given visitor.

visit_hue

Visits children of the nodes with the given visitor.

visit_i_32

Visits children of the nodes with the given visitor.

visit_id_selector

Visits children of the nodes with the given visitor.

visit_ident

Visits children of the nodes with the given visitor.

visit_idents

Visits children of the nodes with the given visitor.

visit_import_conditions

Visits children of the nodes with the given visitor.

visit_import_href

Visits children of the nodes with the given visitor.

visit_import_layer_name

Visits children of the nodes with the given visitor.

visit_import_prelude

Visits children of the nodes with the given visitor.

visit_important_flag

Visits children of the nodes with the given visitor.

visit_integer

Visits children of the nodes with the given visitor.

visit_keyframe_block

Visits children of the nodes with the given visitor.

visit_keyframe_selector

Visits children of the nodes with the given visitor.

visit_keyframe_selectors

Visits children of the nodes with the given visitor.

visit_keyframes_name

Visits children of the nodes with the given visitor.

visit_keyframes_pseudo_function

Visits children of the nodes with the given visitor.

visit_keyframes_pseudo_prefix

Visits children of the nodes with the given visitor.

visit_layer_name

Visits children of the nodes with the given visitor.

visit_layer_name_list

Visits children of the nodes with the given visitor.

visit_layer_names

Visits children of the nodes with the given visitor.

visit_layer_prelude

Visits children of the nodes with the given visitor.

visit_length

Visits children of the nodes with the given visitor.

visit_length_percentage

Visits children of the nodes with the given visitor.

visit_list_of_component_values

Visits children of the nodes with the given visitor.

visit_media_and

Visits children of the nodes with the given visitor.

visit_media_condition

Visits children of the nodes with the given visitor.

visit_media_condition_all_type

Visits children of the nodes with the given visitor.

visit_media_condition_all_types

Visits children of the nodes with the given visitor.

visit_media_condition_type

Visits children of the nodes with the given visitor.

visit_media_condition_without_or

Visits children of the nodes with the given visitor.

visit_media_condition_without_or_type

Visits children of the nodes with the given visitor.

visit_media_condition_without_or_types

Visits children of the nodes with the given visitor.

visit_media_feature

Visits children of the nodes with the given visitor.

visit_media_feature_boolean

Visits children of the nodes with the given visitor.

visit_media_feature_name

Visits children of the nodes with the given visitor.

visit_media_feature_plain

Visits children of the nodes with the given visitor.

visit_media_feature_range

Visits children of the nodes with the given visitor.

visit_media_feature_range_comparison

Visits children of the nodes with the given visitor.

visit_media_feature_range_interval

Visits children of the nodes with the given visitor.

visit_media_feature_value

Visits children of the nodes with the given visitor.

visit_media_in_parens

Visits children of the nodes with the given visitor.

visit_media_not

Visits children of the nodes with the given visitor.

visit_media_or

Visits children of the nodes with the given visitor.

visit_media_queries

Visits children of the nodes with the given visitor.

visit_media_query

Visits children of the nodes with the given visitor.

visit_media_query_list

Visits children of the nodes with the given visitor.

visit_media_type

Visits children of the nodes with the given visitor.

visit_mut_absolute_color_base

visit_mut_alpha_value

visit_mut_an_plus_b

visit_mut_an_plus_b_notation

visit_mut_angle

visit_mut_angle_percentage

visit_mut_any_namespace

visit_mut_at_rule

visit_mut_at_rule_name

visit_mut_at_rule_prelude

visit_mut_atom

visit_mut_attribute_selector

visit_mut_attribute_selector_matcher

visit_mut_attribute_selector_matcher_value

visit_mut_attribute_selector_modifier

visit_mut_attribute_selector_value

visit_mut_calc_operator

visit_mut_calc_operator_type

visit_mut_calc_product

visit_mut_calc_product_or_operator

visit_mut_calc_product_or_operators

visit_mut_calc_sum

visit_mut_calc_value

visit_mut_calc_value_or_operator

visit_mut_calc_value_or_operators

visit_mut_class_selector

visit_mut_cmyk_component

visit_mut_color

visit_mut_color_profile_name

visit_mut_combinator

visit_mut_combinator_value

visit_mut_complex_selector

visit_mut_complex_selector_children

visit_mut_complex_selector_children_vec

visit_mut_complex_selectors

visit_mut_component_value

visit_mut_component_values

visit_mut_compound_selector

visit_mut_compound_selector_list

visit_mut_compound_selectors

visit_mut_container_condition

visit_mut_container_name

visit_mut_container_query

visit_mut_container_query_and

visit_mut_container_query_not

visit_mut_container_query_or

visit_mut_container_query_type

visit_mut_container_query_types

visit_mut_custom_highlight_name

visit_mut_custom_ident

visit_mut_custom_idents

visit_mut_custom_media_query

visit_mut_custom_media_query_media_type

visit_mut_custom_property_name

visit_mut_dashed_ident

visit_mut_declaration

visit_mut_declaration_name

visit_mut_declaration_or_at_rule

visit_mut_delimiter

visit_mut_delimiter_value

visit_mut_dimension

visit_mut_document_prelude

visit_mut_document_prelude_matching_function

visit_mut_document_prelude_matching_functions

visit_mut_extension_name

visit_mut_family_name

visit_mut_family_names

visit_mut_flex

visit_mut_font_feature_values_prelude

visit_mut_forgiving_complex_selector

visit_mut_forgiving_complex_selectors

visit_mut_forgiving_relative_selector

visit_mut_forgiving_relative_selector_list

visit_mut_forgiving_relative_selectors

visit_mut_forgiving_selector_list

visit_mut_frequency

visit_mut_frequency_percentage

visit_mut_function

visit_mut_function_name

visit_mut_general_enclosed

visit_mut_hex_color

visit_mut_hue

visit_mut_i_32

visit_mut_id_selector

visit_mut_ident

visit_mut_idents

visit_mut_import_conditions

visit_mut_import_href

visit_mut_import_layer_name

visit_mut_import_prelude

visit_mut_important_flag

visit_mut_integer

visit_mut_keyframe_block

visit_mut_keyframe_selector

visit_mut_keyframe_selectors

visit_mut_keyframes_name

visit_mut_keyframes_pseudo_function

visit_mut_keyframes_pseudo_prefix

visit_mut_layer_name

visit_mut_layer_name_list

visit_mut_layer_names

visit_mut_layer_prelude

visit_mut_length

visit_mut_length_percentage

visit_mut_list_of_component_values

visit_mut_media_and

visit_mut_media_condition

visit_mut_media_condition_all_type

visit_mut_media_condition_all_types

visit_mut_media_condition_type

visit_mut_media_condition_without_or

visit_mut_media_condition_without_or_type

visit_mut_media_condition_without_or_types

visit_mut_media_feature

visit_mut_media_feature_boolean

visit_mut_media_feature_name

visit_mut_media_feature_plain

visit_mut_media_feature_range

visit_mut_media_feature_range_comparison

visit_mut_media_feature_range_interval

visit_mut_media_feature_value

visit_mut_media_in_parens

visit_mut_media_not

visit_mut_media_or

visit_mut_media_queries

visit_mut_media_query

visit_mut_media_query_list

visit_mut_media_type

visit_mut_named_namespace

visit_mut_namespace

visit_mut_namespace_prefix

visit_mut_namespace_prelude

visit_mut_namespace_prelude_uri

visit_mut_nesting_selector

visit_mut_number

visit_mut_opt_at_rule_prelude

visit_mut_opt_atom

visit_mut_opt_attribute_selector_matcher

visit_mut_opt_attribute_selector_modifier

visit_mut_opt_attribute_selector_value

visit_mut_opt_combinator

visit_mut_opt_container_name

visit_mut_opt_forgiving_selector_list

visit_mut_opt_function

visit_mut_opt_i_32

visit_mut_opt_ident

visit_mut_opt_import_conditions

visit_mut_opt_import_layer_name

visit_mut_opt_important_flag

visit_mut_opt_media_condition_type

visit_mut_opt_media_query_list

visit_mut_opt_media_type

visit_mut_opt_namespace

visit_mut_opt_namespace_prefix

visit_mut_opt_nesting_selector

visit_mut_opt_number

visit_mut_opt_page_selector_pseudos

visit_mut_opt_page_selector_type

visit_mut_opt_pseudo_class_selector_children_vec

visit_mut_opt_pseudo_element_selector_children_vec

visit_mut_opt_simple_block

visit_mut_opt_type_selector

visit_mut_opt_url_modifiers

visit_mut_opt_url_value

visit_mut_page_selector

visit_mut_page_selector_list

visit_mut_page_selector_pseudo

visit_mut_page_selector_pseudos

visit_mut_page_selector_type

visit_mut_page_selectors

visit_mut_percentage

visit_mut_pseudo_class_selector

visit_mut_pseudo_class_selector_children

visit_mut_pseudo_class_selector_children_vec

visit_mut_pseudo_element_selector

visit_mut_pseudo_element_selector_children

visit_mut_pseudo_element_selector_children_vec

visit_mut_qualified_rule

visit_mut_qualified_rule_prelude

visit_mut_query_in_parens

visit_mut_ratio

visit_mut_relative_selector

visit_mut_relative_selector_list

visit_mut_relative_selectors

visit_mut_resolution

visit_mut_rule

visit_mut_rules

visit_mut_scope_range

visit_mut_selector_list

visit_mut_sequence_of_custom_idents

visit_mut_simple_block

visit_mut_size_feature

visit_mut_size_feature_boolean

visit_mut_size_feature_name

visit_mut_size_feature_plain

visit_mut_size_feature_range

visit_mut_size_feature_range_comparison

visit_mut_size_feature_range_interval

visit_mut_size_feature_value

visit_mut_span

visit_mut_str

visit_mut_style_block

visit_mut_stylesheet

visit_mut_subclass_selector

visit_mut_subclass_selectors

visit_mut_supports_and

visit_mut_supports_condition

visit_mut_supports_condition_type

visit_mut_supports_condition_types

visit_mut_supports_feature

visit_mut_supports_in_parens

visit_mut_supports_not

visit_mut_supports_or

visit_mut_tag_name_selector

visit_mut_time

visit_mut_time_percentage

visit_mut_token

visit_mut_token_and_span

visit_mut_type_selector

visit_mut_unicode_range

visit_mut_universal_selector

visit_mut_unknown_dimension

visit_mut_url

visit_mut_url_modifier

visit_mut_url_modifiers

visit_mut_url_value

visit_mut_url_value_raw

visit_mut_wq_name

visit_named_namespace

Visits children of the nodes with the given visitor.

visit_namespace

Visits children of the nodes with the given visitor.

visit_namespace_prefix

Visits children of the nodes with the given visitor.

visit_namespace_prelude

Visits children of the nodes with the given visitor.

visit_namespace_prelude_uri

Visits children of the nodes with the given visitor.

visit_nesting_selector

Visits children of the nodes with the given visitor.

visit_number

Visits children of the nodes with the given visitor.

visit_opt_at_rule_prelude

Visits children of the nodes with the given visitor.

visit_opt_atom

Visits children of the nodes with the given visitor.

visit_opt_attribute_selector_matcher

Visits children of the nodes with the given visitor.

visit_opt_attribute_selector_modifier

Visits children of the nodes with the given visitor.

visit_opt_attribute_selector_value

Visits children of the nodes with the given visitor.

visit_opt_combinator

Visits children of the nodes with the given visitor.

visit_opt_container_name

Visits children of the nodes with the given visitor.

visit_opt_forgiving_selector_list

Visits children of the nodes with the given visitor.

visit_opt_function

Visits children of the nodes with the given visitor.

visit_opt_i_32

Visits children of the nodes with the given visitor.

visit_opt_ident

Visits children of the nodes with the given visitor.

visit_opt_import_conditions

Visits children of the nodes with the given visitor.

visit_opt_import_layer_name

Visits children of the nodes with the given visitor.

visit_opt_important_flag

Visits children of the nodes with the given visitor.

visit_opt_media_condition_type

Visits children of the nodes with the given visitor.

visit_opt_media_query_list

Visits children of the nodes with the given visitor.

visit_opt_media_type

Visits children of the nodes with the given visitor.

visit_opt_namespace

Visits children of the nodes with the given visitor.

visit_opt_namespace_prefix

Visits children of the nodes with the given visitor.

visit_opt_nesting_selector

Visits children of the nodes with the given visitor.

visit_opt_number

Visits children of the nodes with the given visitor.

visit_opt_page_selector_pseudos

Visits children of the nodes with the given visitor.

visit_opt_page_selector_type

Visits children of the nodes with the given visitor.

visit_opt_pseudo_class_selector_children_vec

Visits children of the nodes with the given visitor.

visit_opt_pseudo_element_selector_children_vec

Visits children of the nodes with the given visitor.

visit_opt_simple_block

Visits children of the nodes with the given visitor.

visit_opt_type_selector

Visits children of the nodes with the given visitor.

visit_opt_url_modifiers

Visits children of the nodes with the given visitor.

visit_opt_url_value

Visits children of the nodes with the given visitor.

visit_page_selector

Visits children of the nodes with the given visitor.

visit_page_selector_list

Visits children of the nodes with the given visitor.

visit_page_selector_pseudo

Visits children of the nodes with the given visitor.

visit_page_selector_pseudos

Visits children of the nodes with the given visitor.

visit_page_selector_type

Visits children of the nodes with the given visitor.

visit_page_selectors

Visits children of the nodes with the given visitor.

visit_percentage

Visits children of the nodes with the given visitor.

visit_pseudo_class_selector

Visits children of the nodes with the given visitor.

visit_pseudo_class_selector_children

Visits children of the nodes with the given visitor.

visit_pseudo_class_selector_children_vec

Visits children of the nodes with the given visitor.

visit_pseudo_element_selector

Visits children of the nodes with the given visitor.

visit_pseudo_element_selector_children

Visits children of the nodes with the given visitor.

visit_pseudo_element_selector_children_vec

Visits children of the nodes with the given visitor.

visit_qualified_rule

Visits children of the nodes with the given visitor.

visit_qualified_rule_prelude

Visits children of the nodes with the given visitor.

visit_query_in_parens

Visits children of the nodes with the given visitor.

visit_ratio

Visits children of the nodes with the given visitor.

visit_relative_selector

Visits children of the nodes with the given visitor.

visit_relative_selector_list

Visits children of the nodes with the given visitor.

visit_relative_selectors

Visits children of the nodes with the given visitor.

visit_resolution

Visits children of the nodes with the given visitor.

visit_rule

Visits children of the nodes with the given visitor.

visit_rules

Visits children of the nodes with the given visitor.

visit_scope_range

Visits children of the nodes with the given visitor.

visit_selector_list

Visits children of the nodes with the given visitor.

visit_sequence_of_custom_idents

Visits children of the nodes with the given visitor.

visit_simple_block

Visits children of the nodes with the given visitor.

visit_size_feature

Visits children of the nodes with the given visitor.

visit_size_feature_boolean

Visits children of the nodes with the given visitor.

visit_size_feature_name

Visits children of the nodes with the given visitor.

visit_size_feature_plain

Visits children of the nodes with the given visitor.

visit_size_feature_range

Visits children of the nodes with the given visitor.

visit_size_feature_range_comparison

Visits children of the nodes with the given visitor.

visit_size_feature_range_interval

Visits children of the nodes with the given visitor.

visit_size_feature_value

Visits children of the nodes with the given visitor.

visit_span

Visits children of the nodes with the given visitor.

visit_str

Visits children of the nodes with the given visitor.

visit_style_block

Visits children of the nodes with the given visitor.

visit_stylesheet

Visits children of the nodes with the given visitor.

visit_subclass_selector

Visits children of the nodes with the given visitor.

visit_subclass_selectors

Visits children of the nodes with the given visitor.

visit_supports_and

Visits children of the nodes with the given visitor.

visit_supports_condition

Visits children of the nodes with the given visitor.

visit_supports_condition_type

Visits children of the nodes with the given visitor.

visit_supports_condition_types

Visits children of the nodes with the given visitor.

visit_supports_feature

Visits children of the nodes with the given visitor.

visit_supports_in_parens

Visits children of the nodes with the given visitor.

visit_supports_not

Visits children of the nodes with the given visitor.

visit_supports_or

Visits children of the nodes with the given visitor.

visit_tag_name_selector

Visits children of the nodes with the given visitor.

visit_time

Visits children of the nodes with the given visitor.

visit_time_percentage

Visits children of the nodes with the given visitor.

visit_token

Visits children of the nodes with the given visitor.

visit_token_and_span

Visits children of the nodes with the given visitor.

visit_type_selector

Visits children of the nodes with the given visitor.

visit_unicode_range

Visits children of the nodes with the given visitor.

visit_universal_selector

Visits children of the nodes with the given visitor.

visit_unknown_dimension

Visits children of the nodes with the given visitor.

visit_url

Visits children of the nodes with the given visitor.

visit_url_modifier

Visits children of the nodes with the given visitor.

visit_url_modifiers

Visits children of the nodes with the given visitor.

visit_url_value

Visits children of the nodes with the given visitor.

visit_url_value_raw

Visits children of the nodes with the given visitor.

visit_wq_name

Visits children of the nodes with the given visitor.

Type Aliases

AstKindPath

AstNodePath