Trait swc_css_visit::VisitAll

pub trait VisitAll {

    // Provided methods
    fn visit_absolute_color_base(&mut self, n: &AbsoluteColorBase) { ... }
    fn visit_alpha_value(&mut self, n: &AlphaValue) { ... }
    fn visit_an_plus_b(&mut self, n: &AnPlusB) { ... }
    fn visit_an_plus_b_notation(&mut self, n: &AnPlusBNotation) { ... }
    fn visit_angle(&mut self, n: &Angle) { ... }
    fn visit_angle_percentage(&mut self, n: &AnglePercentage) { ... }
    fn visit_any_namespace(&mut self, n: &AnyNamespace) { ... }
    fn visit_at_rule(&mut self, n: &AtRule) { ... }
    fn visit_at_rule_name(&mut self, n: &AtRuleName) { ... }
    fn visit_at_rule_prelude(&mut self, n: &AtRulePrelude) { ... }
    fn visit_atom(&mut self, n: &Atom) { ... }
    fn visit_attribute_selector(&mut self, n: &AttributeSelector) { ... }
    fn visit_attribute_selector_matcher(&mut self, n: &AttributeSelectorMatcher) { ... }
    fn visit_attribute_selector_matcher_value(
        &mut self,
        n: &AttributeSelectorMatcherValue
    ) { ... }
    fn visit_attribute_selector_modifier(
        &mut self,
        n: &AttributeSelectorModifier
    ) { ... }
    fn visit_attribute_selector_value(&mut self, n: &AttributeSelectorValue) { ... }
    fn visit_calc_operator(&mut self, n: &CalcOperator) { ... }
    fn visit_calc_operator_type(&mut self, n: &CalcOperatorType) { ... }
    fn visit_calc_product(&mut self, n: &CalcProduct) { ... }
    fn visit_calc_product_or_operator(&mut self, n: &CalcProductOrOperator) { ... }
    fn visit_calc_product_or_operators(&mut self, n: &[CalcProductOrOperator]) { ... }
    fn visit_calc_sum(&mut self, n: &CalcSum) { ... }
    fn visit_calc_value(&mut self, n: &CalcValue) { ... }
    fn visit_calc_value_or_operator(&mut self, n: &CalcValueOrOperator) { ... }
    fn visit_calc_value_or_operators(&mut self, n: &[CalcValueOrOperator]) { ... }
    fn visit_class_selector(&mut self, n: &ClassSelector) { ... }
    fn visit_cmyk_component(&mut self, n: &CmykComponent) { ... }
    fn visit_color(&mut self, n: &Color) { ... }
    fn visit_color_profile_name(&mut self, n: &ColorProfileName) { ... }
    fn visit_combinator(&mut self, n: &Combinator) { ... }
    fn visit_combinator_value(&mut self, n: &CombinatorValue) { ... }
    fn visit_complex_selector(&mut self, n: &ComplexSelector) { ... }
    fn visit_complex_selector_children(&mut self, n: &ComplexSelectorChildren) { ... }
    fn visit_complex_selector_children_vec(
        &mut self,
        n: &[ComplexSelectorChildren]
    ) { ... }
    fn visit_complex_selectors(&mut self, n: &[ComplexSelector]) { ... }
    fn visit_component_value(&mut self, n: &ComponentValue) { ... }
    fn visit_component_values(&mut self, n: &[ComponentValue]) { ... }
    fn visit_compound_selector(&mut self, n: &CompoundSelector) { ... }
    fn visit_compound_selector_list(&mut self, n: &CompoundSelectorList) { ... }
    fn visit_compound_selectors(&mut self, n: &[CompoundSelector]) { ... }
    fn visit_container_condition(&mut self, n: &ContainerCondition) { ... }
    fn visit_container_name(&mut self, n: &ContainerName) { ... }
    fn visit_container_query(&mut self, n: &ContainerQuery) { ... }
    fn visit_container_query_and(&mut self, n: &ContainerQueryAnd) { ... }
    fn visit_container_query_not(&mut self, n: &ContainerQueryNot) { ... }
    fn visit_container_query_or(&mut self, n: &ContainerQueryOr) { ... }
    fn visit_container_query_type(&mut self, n: &ContainerQueryType) { ... }
    fn visit_container_query_types(&mut self, n: &[ContainerQueryType]) { ... }
    fn visit_custom_highlight_name(&mut self, n: &CustomHighlightName) { ... }
    fn visit_custom_ident(&mut self, n: &CustomIdent) { ... }
    fn visit_custom_idents(&mut self, n: &[CustomIdent]) { ... }
    fn visit_custom_media_query(&mut self, n: &CustomMediaQuery) { ... }
    fn visit_custom_media_query_media_type(
        &mut self,
        n: &CustomMediaQueryMediaType
    ) { ... }
    fn visit_custom_property_name(&mut self, n: &CustomPropertyName) { ... }
    fn visit_dashed_ident(&mut self, n: &DashedIdent) { ... }
    fn visit_declaration(&mut self, n: &Declaration) { ... }
    fn visit_declaration_name(&mut self, n: &DeclarationName) { ... }
    fn visit_declaration_or_at_rule(&mut self, n: &DeclarationOrAtRule) { ... }
    fn visit_delimiter(&mut self, n: &Delimiter) { ... }
    fn visit_delimiter_value(&mut self, n: &DelimiterValue) { ... }
    fn visit_dimension(&mut self, n: &Dimension) { ... }
    fn visit_document_prelude(&mut self, n: &DocumentPrelude) { ... }
    fn visit_document_prelude_matching_function(
        &mut self,
        n: &DocumentPreludeMatchingFunction
    ) { ... }
    fn visit_document_prelude_matching_functions(
        &mut self,
        n: &[DocumentPreludeMatchingFunction]
    ) { ... }
    fn visit_extension_name(&mut self, n: &ExtensionName) { ... }
    fn visit_family_name(&mut self, n: &FamilyName) { ... }
    fn visit_family_names(&mut self, n: &[FamilyName]) { ... }
    fn visit_flex(&mut self, n: &Flex) { ... }
    fn visit_font_feature_values_prelude(
        &mut self,
        n: &FontFeatureValuesPrelude
    ) { ... }
    fn visit_forgiving_complex_selector(&mut self, n: &ForgivingComplexSelector) { ... }
    fn visit_forgiving_complex_selectors(
        &mut self,
        n: &[ForgivingComplexSelector]
    ) { ... }
    fn visit_forgiving_relative_selector(
        &mut self,
        n: &ForgivingRelativeSelector
    ) { ... }
    fn visit_forgiving_relative_selector_list(
        &mut self,
        n: &ForgivingRelativeSelectorList
    ) { ... }
    fn visit_forgiving_relative_selectors(
        &mut self,
        n: &[ForgivingRelativeSelector]
    ) { ... }
    fn visit_forgiving_selector_list(&mut self, n: &ForgivingSelectorList) { ... }
    fn visit_frequency(&mut self, n: &Frequency) { ... }
    fn visit_frequency_percentage(&mut self, n: &FrequencyPercentage) { ... }
    fn visit_function(&mut self, n: &Function) { ... }
    fn visit_function_name(&mut self, n: &FunctionName) { ... }
    fn visit_general_enclosed(&mut self, n: &GeneralEnclosed) { ... }
    fn visit_hex_color(&mut self, n: &HexColor) { ... }
    fn visit_hue(&mut self, n: &Hue) { ... }
    fn visit_i_32(&mut self, n: &i32) { ... }
    fn visit_id_selector(&mut self, n: &IdSelector) { ... }
    fn visit_ident(&mut self, n: &Ident) { ... }
    fn visit_idents(&mut self, n: &[Ident]) { ... }
    fn visit_import_conditions(&mut self, n: &ImportConditions) { ... }
    fn visit_import_href(&mut self, n: &ImportHref) { ... }
    fn visit_import_layer_name(&mut self, n: &ImportLayerName) { ... }
    fn visit_import_prelude(&mut self, n: &ImportPrelude) { ... }
    fn visit_important_flag(&mut self, n: &ImportantFlag) { ... }
    fn visit_integer(&mut self, n: &Integer) { ... }
    fn visit_keyframe_block(&mut self, n: &KeyframeBlock) { ... }
    fn visit_keyframe_selector(&mut self, n: &KeyframeSelector) { ... }
    fn visit_keyframe_selectors(&mut self, n: &[KeyframeSelector]) { ... }
    fn visit_keyframes_name(&mut self, n: &KeyframesName) { ... }
    fn visit_keyframes_pseudo_function(&mut self, n: &KeyframesPseudoFunction) { ... }
    fn visit_keyframes_pseudo_prefix(&mut self, n: &KeyframesPseudoPrefix) { ... }
    fn visit_layer_name(&mut self, n: &LayerName) { ... }
    fn visit_layer_name_list(&mut self, n: &LayerNameList) { ... }
    fn visit_layer_names(&mut self, n: &[LayerName]) { ... }
    fn visit_layer_prelude(&mut self, n: &LayerPrelude) { ... }
    fn visit_length(&mut self, n: &Length) { ... }
    fn visit_length_percentage(&mut self, n: &LengthPercentage) { ... }
    fn visit_list_of_component_values(&mut self, n: &ListOfComponentValues) { ... }
    fn visit_media_and(&mut self, n: &MediaAnd) { ... }
    fn visit_media_condition(&mut self, n: &MediaCondition) { ... }
    fn visit_media_condition_all_type(&mut self, n: &MediaConditionAllType) { ... }
    fn visit_media_condition_all_types(&mut self, n: &[MediaConditionAllType]) { ... }
    fn visit_media_condition_type(&mut self, n: &MediaConditionType) { ... }
    fn visit_media_condition_without_or(&mut self, n: &MediaConditionWithoutOr) { ... }
    fn visit_media_condition_without_or_type(
        &mut self,
        n: &MediaConditionWithoutOrType
    ) { ... }
    fn visit_media_condition_without_or_types(
        &mut self,
        n: &[MediaConditionWithoutOrType]
    ) { ... }
    fn visit_media_feature(&mut self, n: &MediaFeature) { ... }
    fn visit_media_feature_boolean(&mut self, n: &MediaFeatureBoolean) { ... }
    fn visit_media_feature_name(&mut self, n: &MediaFeatureName) { ... }
    fn visit_media_feature_plain(&mut self, n: &MediaFeaturePlain) { ... }
    fn visit_media_feature_range(&mut self, n: &MediaFeatureRange) { ... }
    fn visit_media_feature_range_comparison(
        &mut self,
        n: &MediaFeatureRangeComparison
    ) { ... }
    fn visit_media_feature_range_interval(
        &mut self,
        n: &MediaFeatureRangeInterval
    ) { ... }
    fn visit_media_feature_value(&mut self, n: &MediaFeatureValue) { ... }
    fn visit_media_in_parens(&mut self, n: &MediaInParens) { ... }
    fn visit_media_not(&mut self, n: &MediaNot) { ... }
    fn visit_media_or(&mut self, n: &MediaOr) { ... }
    fn visit_media_queries(&mut self, n: &[MediaQuery]) { ... }
    fn visit_media_query(&mut self, n: &MediaQuery) { ... }
    fn visit_media_query_list(&mut self, n: &MediaQueryList) { ... }
    fn visit_media_type(&mut self, n: &MediaType) { ... }
    fn visit_named_namespace(&mut self, n: &NamedNamespace) { ... }
    fn visit_namespace(&mut self, n: &Namespace) { ... }
    fn visit_namespace_prefix(&mut self, n: &NamespacePrefix) { ... }
    fn visit_namespace_prelude(&mut self, n: &NamespacePrelude) { ... }
    fn visit_namespace_prelude_uri(&mut self, n: &NamespacePreludeUri) { ... }
    fn visit_nesting_selector(&mut self, n: &NestingSelector) { ... }
    fn visit_number(&mut self, n: &Number) { ... }
    fn visit_opt_at_rule_prelude(&mut self, n: Option<&Box<AtRulePrelude>>) { ... }
    fn visit_opt_atom(&mut self, n: Option<&Atom>) { ... }
    fn visit_opt_attribute_selector_matcher(
        &mut self,
        n: Option<&AttributeSelectorMatcher>
    ) { ... }
    fn visit_opt_attribute_selector_modifier(
        &mut self,
        n: Option<&AttributeSelectorModifier>
    ) { ... }
    fn visit_opt_attribute_selector_value(
        &mut self,
        n: Option<&AttributeSelectorValue>
    ) { ... }
    fn visit_opt_combinator(&mut self, n: Option<&Combinator>) { ... }
    fn visit_opt_container_name(&mut self, n: Option<&ContainerName>) { ... }
    fn visit_opt_forgiving_selector_list(
        &mut self,
        n: Option<&ForgivingSelectorList>
    ) { ... }
    fn visit_opt_function(&mut self, n: Option<&Box<Function>>) { ... }
    fn visit_opt_i_32(&mut self, n: Option<&i32>) { ... }
    fn visit_opt_ident(&mut self, n: Option<&Ident>) { ... }
    fn visit_opt_import_conditions(&mut self, n: Option<&Box<ImportConditions>>) { ... }
    fn visit_opt_import_layer_name(&mut self, n: Option<&Box<ImportLayerName>>) { ... }
    fn visit_opt_important_flag(&mut self, n: Option<&ImportantFlag>) { ... }
    fn visit_opt_media_condition_type(
        &mut self,
        n: Option<&Box<MediaConditionType>>
    ) { ... }
    fn visit_opt_media_query_list(&mut self, n: Option<&Box<MediaQueryList>>) { ... }
    fn visit_opt_media_type(&mut self, n: Option<&MediaType>) { ... }
    fn visit_opt_namespace(&mut self, n: Option<&Namespace>) { ... }
    fn visit_opt_namespace_prefix(&mut self, n: Option<&NamespacePrefix>) { ... }
    fn visit_opt_nesting_selector(&mut self, n: Option<&NestingSelector>) { ... }
    fn visit_opt_number(&mut self, n: Option<&Number>) { ... }
    fn visit_opt_page_selector_pseudos(
        &mut self,
        n: Option<&[PageSelectorPseudo]>
    ) { ... }
    fn visit_opt_page_selector_type(&mut self, n: Option<&PageSelectorType>) { ... }
    fn visit_opt_pseudo_class_selector_children_vec(
        &mut self,
        n: Option<&[PseudoClassSelectorChildren]>
    ) { ... }
    fn visit_opt_pseudo_element_selector_children_vec(
        &mut self,
        n: Option<&[PseudoElementSelectorChildren]>
    ) { ... }
    fn visit_opt_simple_block(&mut self, n: Option<&SimpleBlock>) { ... }
    fn visit_opt_type_selector(&mut self, n: Option<&Box<TypeSelector>>) { ... }
    fn visit_opt_url_modifiers(&mut self, n: Option<&[UrlModifier]>) { ... }
    fn visit_opt_url_value(&mut self, n: Option<&Box<UrlValue>>) { ... }
    fn visit_page_selector(&mut self, n: &PageSelector) { ... }
    fn visit_page_selector_list(&mut self, n: &PageSelectorList) { ... }
    fn visit_page_selector_pseudo(&mut self, n: &PageSelectorPseudo) { ... }
    fn visit_page_selector_pseudos(&mut self, n: &[PageSelectorPseudo]) { ... }
    fn visit_page_selector_type(&mut self, n: &PageSelectorType) { ... }
    fn visit_page_selectors(&mut self, n: &[PageSelector]) { ... }
    fn visit_percentage(&mut self, n: &Percentage) { ... }
    fn visit_pseudo_class_selector(&mut self, n: &PseudoClassSelector) { ... }
    fn visit_pseudo_class_selector_children(
        &mut self,
        n: &PseudoClassSelectorChildren
    ) { ... }
    fn visit_pseudo_class_selector_children_vec(
        &mut self,
        n: &[PseudoClassSelectorChildren]
    ) { ... }
    fn visit_pseudo_element_selector(&mut self, n: &PseudoElementSelector) { ... }
    fn visit_pseudo_element_selector_children(
        &mut self,
        n: &PseudoElementSelectorChildren
    ) { ... }
    fn visit_pseudo_element_selector_children_vec(
        &mut self,
        n: &[PseudoElementSelectorChildren]
    ) { ... }
    fn visit_qualified_rule(&mut self, n: &QualifiedRule) { ... }
    fn visit_qualified_rule_prelude(&mut self, n: &QualifiedRulePrelude) { ... }
    fn visit_query_in_parens(&mut self, n: &QueryInParens) { ... }
    fn visit_ratio(&mut self, n: &Ratio) { ... }
    fn visit_relative_selector(&mut self, n: &RelativeSelector) { ... }
    fn visit_relative_selector_list(&mut self, n: &RelativeSelectorList) { ... }
    fn visit_relative_selectors(&mut self, n: &[RelativeSelector]) { ... }
    fn visit_resolution(&mut self, n: &Resolution) { ... }
    fn visit_rule(&mut self, n: &Rule) { ... }
    fn visit_rules(&mut self, n: &[Rule]) { ... }
    fn visit_scope_range(&mut self, n: &ScopeRange) { ... }
    fn visit_selector_list(&mut self, n: &SelectorList) { ... }
    fn visit_sequence_of_custom_idents(&mut self, n: &SequenceOfCustomIdents) { ... }
    fn visit_simple_block(&mut self, n: &SimpleBlock) { ... }
    fn visit_size_feature(&mut self, n: &SizeFeature) { ... }
    fn visit_size_feature_boolean(&mut self, n: &SizeFeatureBoolean) { ... }
    fn visit_size_feature_name(&mut self, n: &SizeFeatureName) { ... }
    fn visit_size_feature_plain(&mut self, n: &SizeFeaturePlain) { ... }
    fn visit_size_feature_range(&mut self, n: &SizeFeatureRange) { ... }
    fn visit_size_feature_range_comparison(
        &mut self,
        n: &SizeFeatureRangeComparison
    ) { ... }
    fn visit_size_feature_range_interval(
        &mut self,
        n: &SizeFeatureRangeInterval
    ) { ... }
    fn visit_size_feature_value(&mut self, n: &SizeFeatureValue) { ... }
    fn visit_span(&mut self, n: &Span) { ... }
    fn visit_str(&mut self, n: &Str) { ... }
    fn visit_style_block(&mut self, n: &StyleBlock) { ... }
    fn visit_stylesheet(&mut self, n: &Stylesheet) { ... }
    fn visit_subclass_selector(&mut self, n: &SubclassSelector) { ... }
    fn visit_subclass_selectors(&mut self, n: &[SubclassSelector]) { ... }
    fn visit_supports_and(&mut self, n: &SupportsAnd) { ... }
    fn visit_supports_condition(&mut self, n: &SupportsCondition) { ... }
    fn visit_supports_condition_type(&mut self, n: &SupportsConditionType) { ... }
    fn visit_supports_condition_types(&mut self, n: &[SupportsConditionType]) { ... }
    fn visit_supports_feature(&mut self, n: &SupportsFeature) { ... }
    fn visit_supports_in_parens(&mut self, n: &SupportsInParens) { ... }
    fn visit_supports_not(&mut self, n: &SupportsNot) { ... }
    fn visit_supports_or(&mut self, n: &SupportsOr) { ... }
    fn visit_tag_name_selector(&mut self, n: &TagNameSelector) { ... }
    fn visit_time(&mut self, n: &Time) { ... }
    fn visit_time_percentage(&mut self, n: &TimePercentage) { ... }
    fn visit_token(&mut self, n: &Token) { ... }
    fn visit_token_and_span(&mut self, n: &TokenAndSpan) { ... }
    fn visit_type_selector(&mut self, n: &TypeSelector) { ... }
    fn visit_unicode_range(&mut self, n: &UnicodeRange) { ... }
    fn visit_universal_selector(&mut self, n: &UniversalSelector) { ... }
    fn visit_unknown_dimension(&mut self, n: &UnknownDimension) { ... }
    fn visit_url(&mut self, n: &Url) { ... }
    fn visit_url_modifier(&mut self, n: &UrlModifier) { ... }
    fn visit_url_modifiers(&mut self, n: &[UrlModifier]) { ... }
    fn visit_url_value(&mut self, n: &UrlValue) { ... }
    fn visit_url_value_raw(&mut self, n: &UrlValueRaw) { ... }
    fn visit_wq_name(&mut self, n: &WqName) { ... }

}

Provided Methods

fn visit_absolute_color_base(&mut self, n: &AbsoluteColorBase)

This method can be overridden to customize the visitor behavior.

fn visit_alpha_value(&mut self, n: &AlphaValue)

This method can be overridden to customize the visitor behavior.

fn visit_an_plus_b(&mut self, n: &AnPlusB)

This method can be overridden to customize the visitor behavior.

fn visit_an_plus_b_notation(&mut self, n: &AnPlusBNotation)

This method can be overridden to customize the visitor behavior.

fn visit_angle(&mut self, n: &Angle)

This method can be overridden to customize the visitor behavior.

fn visit_angle_percentage(&mut self, n: &AnglePercentage)

This method can be overridden to customize the visitor behavior.

fn visit_any_namespace(&mut self, n: &AnyNamespace)

This method can be overridden to customize the visitor behavior.

fn visit_at_rule(&mut self, n: &AtRule)

This method can be overridden to customize the visitor behavior.

fn visit_at_rule_name(&mut self, n: &AtRuleName)

This method can be overridden to customize the visitor behavior.

fn visit_at_rule_prelude(&mut self, n: &AtRulePrelude)

This method can be overridden to customize the visitor behavior.

fn visit_atom(&mut self, n: &Atom)

fn visit_attribute_selector(&mut self, n: &AttributeSelector)

This method can be overridden to customize the visitor behavior.

fn visit_attribute_selector_matcher(&mut self, n: &AttributeSelectorMatcher)

This method can be overridden to customize the visitor behavior.

fn visit_attribute_selector_matcher_value( &mut self, n: &AttributeSelectorMatcherValue )

fn visit_attribute_selector_modifier(&mut self, n: &AttributeSelectorModifier)

This method can be overridden to customize the visitor behavior.

fn visit_attribute_selector_value(&mut self, n: &AttributeSelectorValue)

This method can be overridden to customize the visitor behavior.

fn visit_calc_operator(&mut self, n: &CalcOperator)

This method can be overridden to customize the visitor behavior.

fn visit_calc_operator_type(&mut self, n: &CalcOperatorType)

This method can be overridden to customize the visitor behavior.

fn visit_calc_product(&mut self, n: &CalcProduct)

This method can be overridden to customize the visitor behavior.

fn visit_calc_product_or_operator(&mut self, n: &CalcProductOrOperator)

This method can be overridden to customize the visitor behavior.

fn visit_calc_product_or_operators(&mut self, n: &[CalcProductOrOperator])

fn visit_calc_sum(&mut self, n: &CalcSum)

This method can be overridden to customize the visitor behavior.

fn visit_calc_value(&mut self, n: &CalcValue)

This method can be overridden to customize the visitor behavior.

fn visit_calc_value_or_operator(&mut self, n: &CalcValueOrOperator)

This method can be overridden to customize the visitor behavior.

fn visit_calc_value_or_operators(&mut self, n: &[CalcValueOrOperator])

fn visit_class_selector(&mut self, n: &ClassSelector)

This method can be overridden to customize the visitor behavior.

fn visit_cmyk_component(&mut self, n: &CmykComponent)

This method can be overridden to customize the visitor behavior.

fn visit_color(&mut self, n: &Color)

This method can be overridden to customize the visitor behavior.

fn visit_color_profile_name(&mut self, n: &ColorProfileName)

This method can be overridden to customize the visitor behavior.

fn visit_combinator(&mut self, n: &Combinator)

This method can be overridden to customize the visitor behavior.

fn visit_combinator_value(&mut self, n: &CombinatorValue)

fn visit_complex_selector(&mut self, n: &ComplexSelector)

This method can be overridden to customize the visitor behavior.

fn visit_complex_selector_children(&mut self, n: &ComplexSelectorChildren)

This method can be overridden to customize the visitor behavior.

fn visit_complex_selector_children_vec(&mut self, n: &[ComplexSelectorChildren])

fn visit_complex_selectors(&mut self, n: &[ComplexSelector])

fn visit_component_value(&mut self, n: &ComponentValue)

This method can be overridden to customize the visitor behavior.

fn visit_component_values(&mut self, n: &[ComponentValue])

fn visit_compound_selector(&mut self, n: &CompoundSelector)

This method can be overridden to customize the visitor behavior.

fn visit_compound_selector_list(&mut self, n: &CompoundSelectorList)

This method can be overridden to customize the visitor behavior.

fn visit_compound_selectors(&mut self, n: &[CompoundSelector])

fn visit_container_condition(&mut self, n: &ContainerCondition)

This method can be overridden to customize the visitor behavior.

fn visit_container_name(&mut self, n: &ContainerName)

This method can be overridden to customize the visitor behavior.

fn visit_container_query(&mut self, n: &ContainerQuery)

This method can be overridden to customize the visitor behavior.

fn visit_container_query_and(&mut self, n: &ContainerQueryAnd)

This method can be overridden to customize the visitor behavior.

fn visit_container_query_not(&mut self, n: &ContainerQueryNot)

This method can be overridden to customize the visitor behavior.

fn visit_container_query_or(&mut self, n: &ContainerQueryOr)

This method can be overridden to customize the visitor behavior.

fn visit_container_query_type(&mut self, n: &ContainerQueryType)

This method can be overridden to customize the visitor behavior.

fn visit_container_query_types(&mut self, n: &[ContainerQueryType])

fn visit_custom_highlight_name(&mut self, n: &CustomHighlightName)

This method can be overridden to customize the visitor behavior.

fn visit_custom_ident(&mut self, n: &CustomIdent)

This method can be overridden to customize the visitor behavior.

fn visit_custom_idents(&mut self, n: &[CustomIdent])

fn visit_custom_media_query(&mut self, n: &CustomMediaQuery)

This method can be overridden to customize the visitor behavior.

fn visit_custom_media_query_media_type(&mut self, n: &CustomMediaQueryMediaType)

This method can be overridden to customize the visitor behavior.

fn visit_custom_property_name(&mut self, n: &CustomPropertyName)

This method can be overridden to customize the visitor behavior.

fn visit_dashed_ident(&mut self, n: &DashedIdent)

This method can be overridden to customize the visitor behavior.

fn visit_declaration(&mut self, n: &Declaration)

This method can be overridden to customize the visitor behavior.

fn visit_declaration_name(&mut self, n: &DeclarationName)

This method can be overridden to customize the visitor behavior.

fn visit_declaration_or_at_rule(&mut self, n: &DeclarationOrAtRule)

This method can be overridden to customize the visitor behavior.

fn visit_delimiter(&mut self, n: &Delimiter)

This method can be overridden to customize the visitor behavior.

fn visit_delimiter_value(&mut self, n: &DelimiterValue)

This method can be overridden to customize the visitor behavior.

fn visit_dimension(&mut self, n: &Dimension)

This method can be overridden to customize the visitor behavior.

fn visit_document_prelude(&mut self, n: &DocumentPrelude)

This method can be overridden to customize the visitor behavior.

fn visit_document_prelude_matching_function( &mut self, n: &DocumentPreludeMatchingFunction )

This method can be overridden to customize the visitor behavior.

fn visit_document_prelude_matching_functions( &mut self, n: &[DocumentPreludeMatchingFunction] )

fn visit_extension_name(&mut self, n: &ExtensionName)

This method can be overridden to customize the visitor behavior.

fn visit_family_name(&mut self, n: &FamilyName)

This method can be overridden to customize the visitor behavior.

fn visit_family_names(&mut self, n: &[FamilyName])

fn visit_flex(&mut self, n: &Flex)

This method can be overridden to customize the visitor behavior.

fn visit_font_feature_values_prelude(&mut self, n: &FontFeatureValuesPrelude)

This method can be overridden to customize the visitor behavior.

fn visit_forgiving_complex_selector(&mut self, n: &ForgivingComplexSelector)

This method can be overridden to customize the visitor behavior.

fn visit_forgiving_complex_selectors(&mut self, n: &[ForgivingComplexSelector])

fn visit_forgiving_relative_selector(&mut self, n: &ForgivingRelativeSelector)

This method can be overridden to customize the visitor behavior.

fn visit_forgiving_relative_selector_list( &mut self, n: &ForgivingRelativeSelectorList )

This method can be overridden to customize the visitor behavior.

fn visit_forgiving_relative_selectors( &mut self, n: &[ForgivingRelativeSelector] )

fn visit_forgiving_selector_list(&mut self, n: &ForgivingSelectorList)

This method can be overridden to customize the visitor behavior.

fn visit_frequency(&mut self, n: &Frequency)

This method can be overridden to customize the visitor behavior.

fn visit_frequency_percentage(&mut self, n: &FrequencyPercentage)

This method can be overridden to customize the visitor behavior.

fn visit_function(&mut self, n: &Function)

This method can be overridden to customize the visitor behavior.

fn visit_function_name(&mut self, n: &FunctionName)

This method can be overridden to customize the visitor behavior.

fn visit_general_enclosed(&mut self, n: &GeneralEnclosed)

This method can be overridden to customize the visitor behavior.

fn visit_hex_color(&mut self, n: &HexColor)

This method can be overridden to customize the visitor behavior.

fn visit_hue(&mut self, n: &Hue)

This method can be overridden to customize the visitor behavior.

fn visit_i_32(&mut self, n: &i32)

fn visit_id_selector(&mut self, n: &IdSelector)

This method can be overridden to customize the visitor behavior.

fn visit_ident(&mut self, n: &Ident)

This method can be overridden to customize the visitor behavior.

fn visit_idents(&mut self, n: &[Ident])

fn visit_import_conditions(&mut self, n: &ImportConditions)

This method can be overridden to customize the visitor behavior.

fn visit_import_href(&mut self, n: &ImportHref)

This method can be overridden to customize the visitor behavior.

fn visit_import_layer_name(&mut self, n: &ImportLayerName)

This method can be overridden to customize the visitor behavior.

fn visit_import_prelude(&mut self, n: &ImportPrelude)

This method can be overridden to customize the visitor behavior.

fn visit_important_flag(&mut self, n: &ImportantFlag)

This method can be overridden to customize the visitor behavior.

fn visit_integer(&mut self, n: &Integer)

This method can be overridden to customize the visitor behavior.

fn visit_keyframe_block(&mut self, n: &KeyframeBlock)

This method can be overridden to customize the visitor behavior.

fn visit_keyframe_selector(&mut self, n: &KeyframeSelector)

This method can be overridden to customize the visitor behavior.

fn visit_keyframe_selectors(&mut self, n: &[KeyframeSelector])

fn visit_keyframes_name(&mut self, n: &KeyframesName)

This method can be overridden to customize the visitor behavior.

fn visit_keyframes_pseudo_function(&mut self, n: &KeyframesPseudoFunction)

This method can be overridden to customize the visitor behavior.

fn visit_keyframes_pseudo_prefix(&mut self, n: &KeyframesPseudoPrefix)

This method can be overridden to customize the visitor behavior.

fn visit_layer_name(&mut self, n: &LayerName)

This method can be overridden to customize the visitor behavior.

fn visit_layer_name_list(&mut self, n: &LayerNameList)

This method can be overridden to customize the visitor behavior.

fn visit_layer_names(&mut self, n: &[LayerName])

fn visit_layer_prelude(&mut self, n: &LayerPrelude)

This method can be overridden to customize the visitor behavior.

fn visit_length(&mut self, n: &Length)

This method can be overridden to customize the visitor behavior.

fn visit_length_percentage(&mut self, n: &LengthPercentage)

This method can be overridden to customize the visitor behavior.

fn visit_list_of_component_values(&mut self, n: &ListOfComponentValues)

This method can be overridden to customize the visitor behavior.

fn visit_media_and(&mut self, n: &MediaAnd)

This method can be overridden to customize the visitor behavior.

fn visit_media_condition(&mut self, n: &MediaCondition)

This method can be overridden to customize the visitor behavior.

fn visit_media_condition_all_type(&mut self, n: &MediaConditionAllType)

This method can be overridden to customize the visitor behavior.

fn visit_media_condition_all_types(&mut self, n: &[MediaConditionAllType])

fn visit_media_condition_type(&mut self, n: &MediaConditionType)

This method can be overridden to customize the visitor behavior.

fn visit_media_condition_without_or(&mut self, n: &MediaConditionWithoutOr)

This method can be overridden to customize the visitor behavior.

fn visit_media_condition_without_or_type( &mut self, n: &MediaConditionWithoutOrType )

This method can be overridden to customize the visitor behavior.

fn visit_media_condition_without_or_types( &mut self, n: &[MediaConditionWithoutOrType] )

fn visit_media_feature(&mut self, n: &MediaFeature)

This method can be overridden to customize the visitor behavior.

fn visit_media_feature_boolean(&mut self, n: &MediaFeatureBoolean)

This method can be overridden to customize the visitor behavior.

fn visit_media_feature_name(&mut self, n: &MediaFeatureName)

This method can be overridden to customize the visitor behavior.

fn visit_media_feature_plain(&mut self, n: &MediaFeaturePlain)

This method can be overridden to customize the visitor behavior.

fn visit_media_feature_range(&mut self, n: &MediaFeatureRange)

This method can be overridden to customize the visitor behavior.

fn visit_media_feature_range_comparison( &mut self, n: &MediaFeatureRangeComparison )

fn visit_media_feature_range_interval(&mut self, n: &MediaFeatureRangeInterval)

This method can be overridden to customize the visitor behavior.

fn visit_media_feature_value(&mut self, n: &MediaFeatureValue)

This method can be overridden to customize the visitor behavior.

fn visit_media_in_parens(&mut self, n: &MediaInParens)

This method can be overridden to customize the visitor behavior.

fn visit_media_not(&mut self, n: &MediaNot)

This method can be overridden to customize the visitor behavior.

fn visit_media_or(&mut self, n: &MediaOr)

This method can be overridden to customize the visitor behavior.

fn visit_media_queries(&mut self, n: &[MediaQuery])

fn visit_media_query(&mut self, n: &MediaQuery)

This method can be overridden to customize the visitor behavior.

fn visit_media_query_list(&mut self, n: &MediaQueryList)

This method can be overridden to customize the visitor behavior.

fn visit_media_type(&mut self, n: &MediaType)

This method can be overridden to customize the visitor behavior.

fn visit_named_namespace(&mut self, n: &NamedNamespace)

This method can be overridden to customize the visitor behavior.

fn visit_namespace(&mut self, n: &Namespace)

This method can be overridden to customize the visitor behavior.

fn visit_namespace_prefix(&mut self, n: &NamespacePrefix)

This method can be overridden to customize the visitor behavior.

fn visit_namespace_prelude(&mut self, n: &NamespacePrelude)

This method can be overridden to customize the visitor behavior.

fn visit_namespace_prelude_uri(&mut self, n: &NamespacePreludeUri)

This method can be overridden to customize the visitor behavior.

fn visit_nesting_selector(&mut self, n: &NestingSelector)

This method can be overridden to customize the visitor behavior.

fn visit_number(&mut self, n: &Number)

This method can be overridden to customize the visitor behavior.

fn visit_opt_at_rule_prelude(&mut self, n: Option<&Box<AtRulePrelude>>)

fn visit_opt_atom(&mut self, n: Option<&Atom>)

fn visit_opt_attribute_selector_matcher( &mut self, n: Option<&AttributeSelectorMatcher> )

fn visit_opt_attribute_selector_modifier( &mut self, n: Option<&AttributeSelectorModifier> )

fn visit_opt_attribute_selector_value( &mut self, n: Option<&AttributeSelectorValue> )

fn visit_opt_combinator(&mut self, n: Option<&Combinator>)

fn visit_opt_container_name(&mut self, n: Option<&ContainerName>)

fn visit_opt_forgiving_selector_list( &mut self, n: Option<&ForgivingSelectorList> )

fn visit_opt_function(&mut self, n: Option<&Box<Function>>)

fn visit_opt_i_32(&mut self, n: Option<&i32>)

fn visit_opt_ident(&mut self, n: Option<&Ident>)

fn visit_opt_import_conditions(&mut self, n: Option<&Box<ImportConditions>>)

fn visit_opt_import_layer_name(&mut self, n: Option<&Box<ImportLayerName>>)

fn visit_opt_important_flag(&mut self, n: Option<&ImportantFlag>)

fn visit_opt_media_condition_type( &mut self, n: Option<&Box<MediaConditionType>> )

fn visit_opt_media_query_list(&mut self, n: Option<&Box<MediaQueryList>>)

fn visit_opt_media_type(&mut self, n: Option<&MediaType>)

fn visit_opt_namespace(&mut self, n: Option<&Namespace>)

fn visit_opt_namespace_prefix(&mut self, n: Option<&NamespacePrefix>)

fn visit_opt_nesting_selector(&mut self, n: Option<&NestingSelector>)

fn visit_opt_number(&mut self, n: Option<&Number>)

fn visit_opt_page_selector_pseudos(&mut self, n: Option<&[PageSelectorPseudo]>)

fn visit_opt_page_selector_type(&mut self, n: Option<&PageSelectorType>)

fn visit_opt_pseudo_class_selector_children_vec( &mut self, n: Option<&[PseudoClassSelectorChildren]> )

fn visit_opt_pseudo_element_selector_children_vec( &mut self, n: Option<&[PseudoElementSelectorChildren]> )

fn visit_opt_simple_block(&mut self, n: Option<&SimpleBlock>)

fn visit_opt_type_selector(&mut self, n: Option<&Box<TypeSelector>>)

fn visit_opt_url_modifiers(&mut self, n: Option<&[UrlModifier]>)

fn visit_opt_url_value(&mut self, n: Option<&Box<UrlValue>>)

fn visit_page_selector(&mut self, n: &PageSelector)

This method can be overridden to customize the visitor behavior.

fn visit_page_selector_list(&mut self, n: &PageSelectorList)

This method can be overridden to customize the visitor behavior.

fn visit_page_selector_pseudo(&mut self, n: &PageSelectorPseudo)

This method can be overridden to customize the visitor behavior.

fn visit_page_selector_pseudos(&mut self, n: &[PageSelectorPseudo])

fn visit_page_selector_type(&mut self, n: &PageSelectorType)

This method can be overridden to customize the visitor behavior.

fn visit_page_selectors(&mut self, n: &[PageSelector])

fn visit_percentage(&mut self, n: &Percentage)

This method can be overridden to customize the visitor behavior.

fn visit_pseudo_class_selector(&mut self, n: &PseudoClassSelector)

This method can be overridden to customize the visitor behavior.

fn visit_pseudo_class_selector_children( &mut self, n: &PseudoClassSelectorChildren )

This method can be overridden to customize the visitor behavior.

fn visit_pseudo_class_selector_children_vec( &mut self, n: &[PseudoClassSelectorChildren] )

fn visit_pseudo_element_selector(&mut self, n: &PseudoElementSelector)

This method can be overridden to customize the visitor behavior.

fn visit_pseudo_element_selector_children( &mut self, n: &PseudoElementSelectorChildren )

This method can be overridden to customize the visitor behavior.

fn visit_pseudo_element_selector_children_vec( &mut self, n: &[PseudoElementSelectorChildren] )

fn visit_qualified_rule(&mut self, n: &QualifiedRule)

This method can be overridden to customize the visitor behavior.

fn visit_qualified_rule_prelude(&mut self, n: &QualifiedRulePrelude)

This method can be overridden to customize the visitor behavior.

fn visit_query_in_parens(&mut self, n: &QueryInParens)

This method can be overridden to customize the visitor behavior.

fn visit_ratio(&mut self, n: &Ratio)

This method can be overridden to customize the visitor behavior.

fn visit_relative_selector(&mut self, n: &RelativeSelector)

This method can be overridden to customize the visitor behavior.

fn visit_relative_selector_list(&mut self, n: &RelativeSelectorList)

This method can be overridden to customize the visitor behavior.

fn visit_relative_selectors(&mut self, n: &[RelativeSelector])

fn visit_resolution(&mut self, n: &Resolution)

This method can be overridden to customize the visitor behavior.

fn visit_rule(&mut self, n: &Rule)

This method can be overridden to customize the visitor behavior.

fn visit_rules(&mut self, n: &[Rule])

fn visit_scope_range(&mut self, n: &ScopeRange)

This method can be overridden to customize the visitor behavior.

fn visit_selector_list(&mut self, n: &SelectorList)

This method can be overridden to customize the visitor behavior.

fn visit_sequence_of_custom_idents(&mut self, n: &SequenceOfCustomIdents)

This method can be overridden to customize the visitor behavior.

fn visit_simple_block(&mut self, n: &SimpleBlock)

This method can be overridden to customize the visitor behavior.

fn visit_size_feature(&mut self, n: &SizeFeature)

This method can be overridden to customize the visitor behavior.

fn visit_size_feature_boolean(&mut self, n: &SizeFeatureBoolean)

This method can be overridden to customize the visitor behavior.

fn visit_size_feature_name(&mut self, n: &SizeFeatureName)

This method can be overridden to customize the visitor behavior.

fn visit_size_feature_plain(&mut self, n: &SizeFeaturePlain)

This method can be overridden to customize the visitor behavior.

fn visit_size_feature_range(&mut self, n: &SizeFeatureRange)

This method can be overridden to customize the visitor behavior.

fn visit_size_feature_range_comparison( &mut self, n: &SizeFeatureRangeComparison )

This method can be overridden to customize the visitor behavior.

fn visit_size_feature_range_interval(&mut self, n: &SizeFeatureRangeInterval)

This method can be overridden to customize the visitor behavior.

fn visit_size_feature_value(&mut self, n: &SizeFeatureValue)

This method can be overridden to customize the visitor behavior.

fn visit_span(&mut self, n: &Span)

fn visit_str(&mut self, n: &Str)

This method can be overridden to customize the visitor behavior.

fn visit_style_block(&mut self, n: &StyleBlock)

This method can be overridden to customize the visitor behavior.

fn visit_stylesheet(&mut self, n: &Stylesheet)

This method can be overridden to customize the visitor behavior.

fn visit_subclass_selector(&mut self, n: &SubclassSelector)

This method can be overridden to customize the visitor behavior.

fn visit_subclass_selectors(&mut self, n: &[SubclassSelector])

fn visit_supports_and(&mut self, n: &SupportsAnd)

This method can be overridden to customize the visitor behavior.

fn visit_supports_condition(&mut self, n: &SupportsCondition)

This method can be overridden to customize the visitor behavior.

fn visit_supports_condition_type(&mut self, n: &SupportsConditionType)

This method can be overridden to customize the visitor behavior.

fn visit_supports_condition_types(&mut self, n: &[SupportsConditionType])

fn visit_supports_feature(&mut self, n: &SupportsFeature)

This method can be overridden to customize the visitor behavior.

fn visit_supports_in_parens(&mut self, n: &SupportsInParens)

This method can be overridden to customize the visitor behavior.

fn visit_supports_not(&mut self, n: &SupportsNot)

This method can be overridden to customize the visitor behavior.

fn visit_supports_or(&mut self, n: &SupportsOr)

This method can be overridden to customize the visitor behavior.

fn visit_tag_name_selector(&mut self, n: &TagNameSelector)

This method can be overridden to customize the visitor behavior.

fn visit_time(&mut self, n: &Time)

This method can be overridden to customize the visitor behavior.

fn visit_time_percentage(&mut self, n: &TimePercentage)

This method can be overridden to customize the visitor behavior.

fn visit_token(&mut self, n: &Token)

fn visit_token_and_span(&mut self, n: &TokenAndSpan)

This method can be overridden to customize the visitor behavior.

fn visit_type_selector(&mut self, n: &TypeSelector)

This method can be overridden to customize the visitor behavior.

fn visit_unicode_range(&mut self, n: &UnicodeRange)

This method can be overridden to customize the visitor behavior.

fn visit_universal_selector(&mut self, n: &UniversalSelector)

This method can be overridden to customize the visitor behavior.

fn visit_unknown_dimension(&mut self, n: &UnknownDimension)

This method can be overridden to customize the visitor behavior.

fn visit_url(&mut self, n: &Url)

This method can be overridden to customize the visitor behavior.

fn visit_url_modifier(&mut self, n: &UrlModifier)

This method can be overridden to customize the visitor behavior.

fn visit_url_modifiers(&mut self, n: &[UrlModifier])

fn visit_url_value(&mut self, n: &UrlValue)

This method can be overridden to customize the visitor behavior.

fn visit_url_value_raw(&mut self, n: &UrlValueRaw)

This method can be overridden to customize the visitor behavior.

fn visit_wq_name(&mut self, n: &WqName)

This method can be overridden to customize the visitor behavior.